Pull out SimUtils into its own file
